move `fill_missing` to new module `fill`
